requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
beautifulsoup4==4.6.3
blinker==1.4
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3.1
dnslib==0.9.10
Flask==1.0.2
Flask-Assets==0.12
Flask-Bcrypt==0.7.1
Flask-Compress==1.4.0
Flask-Markdown==0.3
Flask-Misaka==0.4.1
Flask-WTF==0.14.2
google-api-core==1.14.0
google-cloud-storage==1.15.1
icalendar==4.0.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10.1
libsass==0.14.5
lxml==4.2.5
Markdown==2.6.11
MarkupSafe==1.0
misaka==2.1.0
paypalrestsdk==1.13.1
psycopg2==2.8.4
pycparser==2.18
pyOpenSSL==18.0.0
python-dateutil==2.8.1
pytz==2018.5
requests==2.21.0
schedule==0.5.0
setuptools==41.6.0
six==1.11.0
squareup
urllib3==1.24.3
webassets==0.12.1
Werkzeug==0.15.5
WTForms==2.2.1

# Async Server and websockets
gevent-websocket==0.10.1
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
psycogreen==1.0.1
ujson==1.35
wsaccel==0.6.2
